+++ api/models/event.py
@@ -1,0 +1,7 @@
+import datetime as dt
+from mongoengine import *
+
+
+class Event(EmbeddedDocument):
+    datetime = DateTimeField(required=True)
+    created_at = DateTimeField(required=True, default=dt.datetime.now())
