+++ api/models/goal.py
@@ -1,0 +1,10 @@
+import datetime as dt
+from mongoengine import *
+
+from .task import Task
+from .log import Log
+
+
+class Goal(Task):
+    logs = SortedListField(EmbeddedDocumentField(Log))
+    name = StringField(required=True)
