+++ api/models/log_attachment.py
@@ -1,0 +1,7 @@
+import datetime as dt
+from mongoengine import *
+
+
+class LogAttachment(EmbeddedDocument):
+    url = URLField(required=True)
+    created_at = DateTimeField(required=True, default=dt.datetime.now())
