+++ api/resources/__init__.py
@@ -1,0 +1,3 @@
+from .auth import *
+from .users import *
+from .growthbooks import *
